Added Areas. Added cover type controller. Added dropdowm list.
