+++ BulkyBookWeb/Areas/Admin/Views/CoverType/Delete.cshtml
@@ -1,0 +1,27 @@
+﻿@model CoverType
+
+<form method="post" asp-action="Delete">
+    <input asp-for="Id" hidden />
+    <div class="border p-3 mt-4">
+        <div class="row pb-2">
+            <h2 class="text-primary">Delete Cover Type</h2>
+            <hr />
+        </div>
+        @*<div asp-validation-summary="All"></div>*@
+        <div class="mb-3">
+            <label asp-for="Name"></label>
+            <input asp-for="Name" class="form-control" disabled/>
+        </div>
+        <button type="submit" class="btn btn-danger" style="width:150px">Delete</button>
+        <a asp-asp-controller="CoverType" asp-action="Index" class="btn btn-secondary" style="width:150px">
+            Back to List
+        </a>
+    </div>
+</form>
+
+@*@section Scripts
+{
+    @{
+        <partial name="_ValidationScriptsPartial" />
+    }
+}*@
