+++ BulkyBookWeb/Areas/Admin/Views/CoverType/Index.cshtml
@@ -1,0 +1,50 @@
+﻿@model IEnumerable<CoverType>
+@{
+    ViewData["Title"] = "Index";
+}
+
+<div class="container p-3">
+    <div class="row pt-4">
+        <div class="col-6">
+            <h2 class="text-primary">Cover Type</h2>
+        </div>
+        <div class="col-6 text-end">
+            <a asp-controller="CoverType" asp-action="Create" class="btn btn-primary">
+                <i class="bi bi-plus-circle"></i> &nbsp; Create New Cover Type
+            </a>
+        </div>
+    </div>
+    <br /><br />
+
+    <table class="table table-bordered table-striped" style="width:100%">
+        <thead>
+            <tr>
+                <th>
+                    Cover Type Name
+                </th>
+                <th>
+
+                </th>
+            </tr>
+        </thead>
+        <tbody>
+            @foreach(var obj in Model)
+            {
+                <tr>
+                    <td width="50%">
+                        @obj.Name
+                    </td>
+                    <td>
+                        <div class="w-150 btn-group" role="group">
+                            <a asp-controller="CoverType" asp-action="Edit" asp-route-id="@obj.Id" 
+                            class="btn btn-primary mx-2"><i class="bi bi-pencil-square"></i> &nbsp; Edit</a>
+                            <a asp-controller="CoverType" asp-action="Delete" asp-route-id="@obj.Id"
+                            class="btn btn-danger mx-2"><i class="bi bi-x-square"></i> &nbsp; Delete</a>
+                        </div>
+                    </td>
+                </tr>
+            }
+        </tbody>
+    </table>
+
+</div>
