+++ BulkyBookWeb/Areas/Customer/Controllers/HomeController.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using System.Diagnostics;
 
-namespace BulkyBookWeb.Controllers
+namespace BulkyBookWeb.Areas.Customer.Controllers
 {
     public class HomeController : Controller
     {
