+++ BulkyBookWeb/Areas/Customer/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using BulkyBookWeb
+@using BulkyBook.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
